Test cases now verify both 32 and 64 bit binaries.
Defined a "scratch" register macro to use rax or eax, depending on the
architecture.

+++ aux.h
@@ -1,6 +1,6 @@
 #ifndef _AUX_H_
 #define _AUX_H_
 
-void printBinary(int i);
+void printBinary(intptr_t i);
 
 #endif
